+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 - Fixed [Get-Gmail][] Incorrect uri.
+- Updated [Invoke-Link][] Synopsis.
 
 ## [0.12.0] - 2025-07-13